+++ packages/fcl/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 - YYYY-MM-DD **BREAKING?** -- description
+
+- 2021-05-27 -- Updates `fcl.serialize` to return serialized voucher
 - 2021-05-27 -- VSN `@onflow/sdk` 0.0.46-alpha.1 -> 0.0.47-alpha.1
 
 ## 0.0.70 - 2021-05-10